+++ src/Public/Get-MaxPathLength.ps1
@@ -45,9 +45,8 @@ function Get-MaxPathLength {
             # Unicode characters always require 32 bits (4 bytes) of memory
             [int]::Parse((getconf PATH_MAX /), [CultureInfo]::InvariantCulture)
         } else {
-            Write-Error $OperatingSystemNotSupportedError `
-                -Category NotImplemented `
-                -ErrorAction Stop
+            $ErrorRecord = New-TerminatingErrorRecord -Message $OperatingSystemNotSupportedError -Category NotImplemented -ErrorId "OperatingSystemNotSupported"
+            $PSCmdlet.ThrowTerminatingError($ErrorRecord)
         }
 
         Write-Output $MaxPathLength