+++ src/Public/Set-PowerState.ps1
@@ -79,9 +79,8 @@ function Set-PowerState {
                 sudo pmset -a hibernatemode $($PowerState -eq "Hibernate" ? 25 : 3)
                 pmset sleepnow
             } else {
-                Write-Error $OperatingSystemNotSupportedError `
-                    -Category NotImplemented `
-                    -ErrorAction Stop
+                $ErrorRecord = New-TerminatingErrorRecord -Message $OperatingSystemNotSupportedError -Category NotImplemented -ErrorId "OperatingSystemNotSupported"
+                $PSCmdlet.ThrowTerminatingError($ErrorRecord)
             }
         }
     }